+++ packages/multichain/src/scope/authorization.test.ts
@@ -25,10 +25,6 @@ const validScopeObject: ExternalScopeObject = {
 };
 
 describe('Scope Authorization', () => {
-  afterEach(() => {
-    jest.resetAllMocks();
-  });
-
   describe('validateAndNormalizeScopes', () => {
     it('validates the scopes', () => {
       try {